Add net/netatalk4
Netatalk is an Apple Filesharing Protocol server. Version 4
re-integrates the AppleTalk ddp support for Traditional Mac OS and
Apple II that had been removed from version 3.

Moreover, it comes with numerous fixes and improvements.

This version supersedes both the netatalk22 and netatalk3 packages,
which will be removed in due time.