regal: Update to 0.29.2
Changes:
v0.29.2
This patch release fixes an issue where the new defer-assignment rule
would sometimes report a false positive when the variable was used
inside of a `with` clause on the next line.

v0.29.1
This patch release fixes an issue where custom (i.e. user-created)
aggregate rules wouldn't work as expected when the condition for a
violation was the absence of aggregated data. This could for example be
a rule that says "at least one rule must be named allow, and it must
have a default assignment to `false`".

v0.29.0
This is a big release, spanning more than a month of development!
Regal v0.29.0 brings new linter rules, performance improvements and new
features to both the linter and the language server.

New rules:
 - defer-assignment
 - walk-no-path
 - rule-assigns-default

Misc improvements and bug fixes.