regress: sync infrastructure tests with reality
When extracting using tar on NetBSD-10.99-amd64, exclusion patterns may
now contain slashes.

When 'bmake' stops due to an error, the error message contains the
targets that were requested.

When 'bmake' runs a long command (more than 1000 bytes), it first writes
the command to a temporary file and then runs that file, to avoid hitting
command line argument limits. This temporary file is included in the
error message of the 'unreadable file' test.