py-hypothesis: updated to 6.119.4
6.119.4

This patch fixes a bug since 6.99.13

6.119.3

Hypothesis collects coverage information during the shrink and explain phases in order to show a more informative error message. On 3.12+, this uses sys.monitoring. This patch improves the performance of coverage collection on 3.12+ by disabling events we don’t need.

6.119.2

This patch refactors some internals to prepare for future work using our IR

6.119.1

This patch migrates some more internals (around generating novel inputs) to the IR layer

6.119.0

This release improves Hypothesis’ handling of ExceptionGroup - it’s now able to detect marker detections if they’re inside a group and attempts to resolve them. Note that this handling is still a work in progress and might not handle edge cases optimally. Please open issues if you encounter any problems or unexpected behavior with it.

6.118.9

Internal refactorings in preparation for upcoming changes.

6.118.8

Internal renamings.

6.118.7

This patch removes some # type: ignore comments following a mypy update.

6.118.6

When Hypothesis replays examples from its test database that it knows were previously fully shrunk it will no longer try to shrink them again.

This should significantly speed up development workflows for slow tests, as the shrinking could contribute a significant delay when rerunning the tests.

In some rare cases this may cause minor reductions in example quality. This was considered an acceptable tradeoff for the improved test runtime.

6.118.5

This patch avoids computing some string representations we won’t need, giving a small speedup.

6.118.4

This patch migrates the optimisation algorithm for targeted property-based testing to our IR layer

6.118.3

This patch adds more type hints to internal Hypothesis code.

6.118.2

This patch migrates the explain phase to our IR layer

6.118.1

This patch updates some internals around how we determine an input is too large to finish generating.

6.118.0

The urls() strategy no longer generates URLs where the port number is 0.

This change is motivated by the idea that the generated URLs should, at least in theory, be possible to fetch. The port number 0 is special; if a server binds to port 0, the kernel will allocate an unused, and non-zero, port instead. That means that it’s not possible for a server to actually be listening on port 0. This motivation is briefly described in the documentation for urls().